Move GROUP_EDGE_ROLES constants
To avoid circular dependencies between the models and other code,
move GROUP_EDGE_ROLES and related constants to grouper.entities.
These constants are still poorly-structured and need further work,
but that's a major refactoring, so defer that for later.

Move another deferred import in get_template_env that doesn't
create a circular dependency and therefore can be a normal import.
