+++ grouper/graph.py
@@ -10,9 +10,10 @@ from sqlalchemy import or_
 from sqlalchemy.orm import aliased
 from sqlalchemy.sql import label, literal
 
+from grouper.entities.group_edge import GROUP_EDGE_ROLES
 from grouper.models.counter import Counter
 from grouper.models.group import Group
-from grouper.models.group_edge import GROUP_EDGE_ROLES, GroupEdge
+from grouper.models.group_edge import GroupEdge
 from grouper.models.group_service_accounts import GroupServiceAccount
 from grouper.models.permission import MappedPermission, Permission
 from grouper.models.permission_map import PermissionMap
