+++ grouper/repositories/permission_grant.py
@@ -3,6 +3,7 @@ from typing import TYPE_CHECKING
 
 from sqlalchemy import or_
 
+from grouper.entities.group_edge import GROUP_EDGE_ROLES
 from grouper.entities.permission_grant import (
     GroupPermissionGrant,
     PermissionGrant,
@@ -10,7 +11,7 @@ from grouper.entities.permission_grant import (
 )
 from grouper.models.base.constants import OBJ_TYPES
 from grouper.models.group import Group
-from grouper.models.group_edge import GROUP_EDGE_ROLES, GroupEdge
+from grouper.models.group_edge import GroupEdge
 from grouper.models.permission import Permission
 from grouper.models.permission_map import PermissionMap
 from grouper.models.service_account import ServiceAccount
