+++ plugins/group_ownership_policy.py
@@ -1,8 +1,9 @@
 from sqlalchemy.sql import label
 
+from grouper.entities.group_edge import GROUP_EDGE_ROLES, OWNER_ROLE_INDICES
 from grouper.models.base.constants import OBJ_TYPES
 from grouper.models.group import Group
-from grouper.models.group_edge import GROUP_EDGE_ROLES, GroupEdge, OWNER_ROLE_INDICES
+from grouper.models.group_edge import GroupEdge
 from grouper.models.user import User
 from grouper.plugin.base import BasePlugin
 from grouper.plugin.exceptions import (
